Some progress on limiting usage of iOS tech
- WiP on isolating Objective C code
- WiP on moving away from Plists in favour of JSON

+++ src/PDSampler.h
@@ -18,8 +18,8 @@ public:
     PDSampler(NSString *pdFilename);
 
     void loadSample(NSString *sampleFilename);
-    void noteOn(unsigned int note);
-    void noteOff(unsigned int note);
+    void noteOn(unsigned long note);
+    void noteOff(unsigned long note);
 };
 
 #endif /* PDSampler_h */
